# src/main/generated/src/main/kotlin/com/example/usj_backend/api/ProfilesApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package com.example.usj_backend.api

import com.example.usj_backend.model.Profile
import com.example.usj_backend.model.ProfileUpdateRequest
import com.example.usj_backend.model.PublicProfile
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface ProfilesApi {

    @Operation(
        tags = ["profiles",],
        summary = "自分のプロフィール取得",
        operationId = "apiProfilesMeGet",
        description = """認証されたユーザーの詳細プロフィールを取得します""",
        responses = [
            ApiResponse(responseCode = "200", description = "プロフィール取得成功", content = [Content(schema = Schema(implementation = Profile::class))]),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/api/profiles/me"],
            produces = ["application/json"]
    )
    fun apiProfilesMeGet(): ResponseEntity<Profile>

    @Operation(
        tags = ["profiles",],
        summary = "プロフィール更新",
        operationId = "apiProfilesMePut",
        description = """認証されたユーザーのプロフィールを更新します""",
        responses = [
            ApiResponse(responseCode = "200", description = "更新成功", content = [Content(schema = Schema(implementation = Profile::class))]),
            ApiResponse(responseCode = "400", description = "入力エラー"),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.PUT],
            value = ["/api/profiles/me"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun apiProfilesMePut(@Parameter(description = "", required = true) @Valid @RequestBody profileUpdateRequest: ProfileUpdateRequest): ResponseEntity<Profile>

    @Operation(
        tags = ["profiles",],
        summary = "他ユーザーのプロフィール取得",
        operationId = "apiProfilesUserIdGet",
        description = """指定したユーザーの公開プロフィールを取得します""",
        responses = [
            ApiResponse(responseCode = "200", description = "プロフィール取得成功", content = [Content(schema = Schema(implementation = PublicProfile::class))]),
            ApiResponse(responseCode = "401", description = "未認証"),
            ApiResponse(responseCode = "404", description = "ユーザーが見つかりません")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/api/profiles/{userId}"],
            produces = ["application/json"]
    )
    fun apiProfilesUserIdGet(@Parameter(description = "", required = true) @PathVariable("userId") userId: kotlin.Long): ResponseEntity<PublicProfile>
}