# src/main/generated/src/main/kotlin/com/example/usj_backend/api/UsersApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package com.example.usj_backend.api

import com.example.usj_backend.model.User
import com.example.usj_backend.model.UserUpdateRequest
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface UsersApi {

    @Operation(
        tags = ["users",],
        summary = "現在のユーザー情報取得",
        operationId = "apiUsersMeGet",
        description = """認証されたユーザーの情報を取得します""",
        responses = [
            ApiResponse(responseCode = "200", description = "ユーザー情報取得成功", content = [Content(schema = Schema(implementation = User::class))]),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/api/users/me"],
            produces = ["application/json"]
    )
    fun apiUsersMeGet(): ResponseEntity<User>

    @Operation(
        tags = ["users",],
        summary = "ユーザー情報更新",
        operationId = "apiUsersMePut",
        description = """認証されたユーザーの基本情報を更新します""",
        responses = [
            ApiResponse(responseCode = "200", description = "更新成功", content = [Content(schema = Schema(implementation = User::class))]),
            ApiResponse(responseCode = "400", description = "入力エラー"),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.PUT],
            value = ["/api/users/me"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun apiUsersMePut(@Parameter(description = "", required = true) @Valid @RequestBody userUpdateRequest: UserUpdateRequest): ResponseEntity<User>
}