# src/main/generated/src/main/kotlin/com/example/usj_backend/model/CreateDatePlanRequest.kt
package com.example.usj_backend.model

import java.util.Objects
import com.example.usj_backend.model.CreateDatePlanItemRequest
import com.fasterxml.jackson.annotation.JsonProperty
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param title 
 * @param scheduledDate 
 * @param description 
 * @param items 
 */
data class CreateDatePlanRequest(

    @get:Size(max=100)
    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("title", required = true) val title: kotlin.String,

    @field:Valid
    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("scheduledDate", required = true) val scheduledDate: java.time.LocalDate,

    @get:Size(max=500)
    @Schema(example = "null", description = "")
    @get:JsonProperty("description") val description: kotlin.String? = null,

    @field:Valid
    @Schema(example = "null", description = "")
    @get:JsonProperty("items") val items: kotlin.collections.List<CreateDatePlanItemRequest>? = null
    ) {

}